refactor(widget): 프리셋 선택 모달을 독립 컴포넌트로 분리
